Move class 'RuleNames' to o.e.xtext.xtext
