+++ intellij/org.eclipse.xtext.idea.generator/xtend-gen/org/eclipse/xtext/idea/generator/parser/antlr/DefaultAntlrGrammarGenerator.java
@@ -29,7 +29,6 @@ import org.eclipse.xtext.Group;
 import org.eclipse.xtext.Keyword;
 import org.eclipse.xtext.ParserRule;
 import org.eclipse.xtext.RuleCall;
-import org.eclipse.xtext.RuleNames;
 import org.eclipse.xtext.TerminalRule;
 import org.eclipse.xtext.UnorderedGroup;
 import org.eclipse.xtext.generator.Naming;
@@ -44,6 +43,7 @@ import org.eclipse.xtext.idea.generator.parser.antlr.XtextIDEAGeneratorExtension
 import org.eclipse.xtext.xbase.lib.Extension;
 import org.eclipse.xtext.xbase.lib.Functions.Function1;
 import org.eclipse.xtext.xbase.lib.IterableExtensions;
+import org.eclipse.xtext.xtext.RuleNames;
 import org.eclipse.xtext.xtext.generator.normalization.FlattenedGrammarAccess;
 import org.eclipse.xtext.xtext.generator.normalization.RuleFilter;
