+++ intellij/org.eclipse.xtext.idea/xtend-gen/org/eclipse/xtext/idea/resource/PsiToEcoreTransformator.java
@@ -33,7 +33,6 @@ import org.eclipse.xtext.GrammarUtil;
 import org.eclipse.xtext.Keyword;
 import org.eclipse.xtext.ParserRule;
 import org.eclipse.xtext.RuleCall;
-import org.eclipse.xtext.RuleNames;
 import org.eclipse.xtext.TerminalRule;
 import org.eclipse.xtext.idea.lang.GrammarAwarePsiErrorElement;
 import org.eclipse.xtext.idea.resource.PsiToEcoreAdapter;
@@ -50,6 +49,7 @@ import org.eclipse.xtext.xbase.lib.Conversions;
 import org.eclipse.xtext.xbase.lib.Extension;
 import org.eclipse.xtext.xbase.lib.IterableExtensions;
 import org.eclipse.xtext.xbase.lib.Pure;
+import org.eclipse.xtext.xtext.RuleNames;
 
 @SuppressWarnings("all")
 public class PsiToEcoreTransformator implements IParser {
