+++ plugins/org.eclipse.xtext.generator/src/org/eclipse/xtext/generator/parser/antlr/AbstractAntlrGeneratorFragment.java
@@ -19,7 +19,6 @@ import org.eclipse.xtext.AbstractRule;
 import org.eclipse.xtext.Grammar;
 import org.eclipse.xtext.GrammarUtil;
 import org.eclipse.xtext.ParserRule;
-import org.eclipse.xtext.RuleNames;
 import org.eclipse.xtext.UnorderedGroup;
 import org.eclipse.xtext.generator.AbstractGeneratorFragment;
 import org.eclipse.xtext.generator.NewlineNormalizer;
@@ -33,6 +32,7 @@ import org.eclipse.xtext.xtext.generator.parser.antlr.splitting.BacktrackingGuar
 import org.eclipse.xtext.xtext.generator.parser.antlr.splitting.PartialClassExtractor;
 import org.eclipse.xtext.xtext.generator.parser.antlr.splitting.SyntacticPredicateFixup;
 import org.eclipse.xtext.xtext.generator.parser.antlr.splitting.UnorderedGroupsSplitter;
+import org.eclipse.xtext.xtext.RuleNames;
 import org.eclipse.xtext.xtext.generator.normalization.FlattenedGrammarAccess;
 import org.eclipse.xtext.xtext.generator.normalization.RuleFilter;
