+++ plugins/org.eclipse.xtext.xtext.generator/src/org/eclipse/xtext/xtext/generator/parser/antlr/AntlrGrammarGenUtil.java
@@ -30,12 +30,12 @@ import org.eclipse.xtext.Parameter;
 import org.eclipse.xtext.ParameterReference;
 import org.eclipse.xtext.ParserRule;
 import org.eclipse.xtext.RuleCall;
-import org.eclipse.xtext.RuleNames;
 import org.eclipse.xtext.XtextFactory;
 import org.eclipse.xtext.util.Strings;
 import org.eclipse.xtext.util.XtextSwitch;
 import org.eclipse.xtext.xbase.lib.Functions;
 import org.eclipse.xtext.xbase.lib.IterableExtensions;
+import org.eclipse.xtext.xtext.RuleNames;
 import org.eclipse.xtext.xtext.generator.normalization.OriginalElement;
 import org.eclipse.xtext.xtext.generator.normalization.OriginalGrammar;
 import org.eclipse.xtext.xtext.generator.normalization.RuleWithParameterValues;
