+++ plugins/org.eclipse.xtext.xtext.generator/xtend-gen/org/eclipse/xtext/xtext/generator/normalization/FlattenedGrammarAccess.java
@@ -42,7 +42,6 @@ import org.eclipse.xtext.NamedArgument;
 import org.eclipse.xtext.Parameter;
 import org.eclipse.xtext.ParserRule;
 import org.eclipse.xtext.RuleCall;
-import org.eclipse.xtext.RuleNames;
 import org.eclipse.xtext.TerminalRule;
 import org.eclipse.xtext.XtextPackage;
 import org.eclipse.xtext.util.internal.EmfAdaptable;
@@ -54,6 +53,7 @@ import org.eclipse.xtext.xbase.lib.Procedures.Procedure1;
 import org.eclipse.xtext.xbase.lib.Procedures.Procedure2;
 import org.eclipse.xtext.xbase.lib.Pure;
 import org.eclipse.xtext.xtext.ConditionEvaluator;
+import org.eclipse.xtext.xtext.RuleNames;
 import org.eclipse.xtext.xtext.UsedRulesFinder;
 import org.eclipse.xtext.xtext.generator.normalization.OriginalElement;
 import org.eclipse.xtext.xtext.generator.normalization.OriginalGrammar;
