+++ tests/org.eclipse.xtext.tests/xtend-gen/org/eclipse/xtext/xtext/generator/GrammarFlatteningTest.java
@@ -12,11 +12,11 @@ import org.eclipse.emf.common.util.URI;
 import org.eclipse.emf.ecore.EObject;
 import org.eclipse.xtend2.lib.StringConcatenation;
 import org.eclipse.xtext.Grammar;
-import org.eclipse.xtext.RuleNames;
 import org.eclipse.xtext.XtextStandaloneSetup;
 import org.eclipse.xtext.junit4.AbstractXtextTests;
 import org.eclipse.xtext.resource.XtextResource;
 import org.eclipse.xtext.serializer.ISerializer;
+import org.eclipse.xtext.xtext.RuleNames;
 import org.eclipse.xtext.xtext.generator.normalization.FlattenedGrammarAccess;
 import org.eclipse.xtext.xtext.generator.normalization.RuleFilter;
 import org.junit.Assert;
